refactor: move client restrictions util out of chats module
